move SyncCapableCommCareActivity to activities pkg
