+++ app/src/org/commcare/activities/PostRequestActivity.java
@@ -10,9 +10,7 @@ import android.widget.TextView;
 import org.commcare.core.network.ModernHttpRequester;
 import org.commcare.dalvik.R;
 import org.commcare.interfaces.ConnectorWithHttpResponseProcessor;
-import org.commcare.interfaces.SyncCapableCommCareActivity;
 import org.commcare.tasks.DataPullTask;
-import org.commcare.tasks.ResultAndError;
 import org.commcare.tasks.SimpleHttpTask;
 import org.commcare.tasks.templates.CommCareTaskConnector;
 import org.commcare.views.ManagedUi;
