+++ app/src/org/commcare/activities/SyncCapableCommCareActivity.java
@@ -1,4 +1,4 @@
-package org.commcare.interfaces;
+package org.commcare.activities;
 
 import android.os.Bundle;
